requirements.txt
streamlit
streamlit-webrtc
opencv-python
dlib-bin
imutils
scipy
pygame
numpy
